05week\checkers.js replaced checkers.js with newCheckers.js
